+++ Sources/VulkanUtility.h
@@ -54,4 +54,4 @@ void GenerateMipmaps(VkPhysicalDevice physicalDevice, VkDevice logicalDevice, Vk
 std::tuple<VkPipeline, VkPipelineLayout> CreateComputePipeline(VkDevice logicalDevice, VkShaderModule computeShaderModule, VkDescriptorSetLayout descriptorSetLayout);
 std::tuple<VkPipeline, VkPipelineLayout> CreateComputePipeline(VkDevice logicalDevice, VkShaderModule computeShaderModule, VkDescriptorSetLayout descriptorSetLayout, const std::vector<VkPushConstantRange> &pushConstantRanges);
 
-size_t DivisionCeil(size_t x, size_t y);
+uint32_t DivisionCeil(uint32_t x, uint32_t y);
